+++ src/python/Components/Lockers/ModuleLocker.py
@@ -9,7 +9,7 @@ from time import struct_time, gmtime, mktime, asctime, sleep
 from typing import Optional, List
 
 from Utilities.Logging import displayTime
-from Services.Mongo.MongoClient import MongoClient
+from Databases.Mongo.MongoClient import MongoClient
 
 
 class ModuleLocker(MongoClient):
