+++ src/python/Components/Monitors/WorkflowCloseoutMonitor.py
@@ -9,10 +9,10 @@ from typing import List, Optional
 
 from Utilities.Logging import displayNumber
 from Utilities.ConfigurationHandler import ConfigurationHandler
-from Services.EOS.EOSWriter import EOSWriter
-from Services.Mongo.MongoClient import MongoClient
-from Services.Oracle.OracleClient import OracleClient
-from Services.Oracle.OracleDB import Workflow
+from Storage.EOS.EOSWriter import EOSWriter
+from Databases.Mongo.MongoClient import MongoClient
+from Databases.Oracle.OracleClient import OracleClient
+from Databases.Oracle.OracleDB import Workflow
 
 
 class WorkflowCloseoutMonitor(MongoClient, OracleClient):
