+++ test/python/Components_t/Monitors_t/WorkflowCloseoutMonitor_t.py
@@ -7,7 +7,7 @@ Unit test for WorkflowCloseoutMonitor helper class.
 import unittest
 from pymongo.collection import Collection
 
-from Services.Mongo.MongoClient import MongoClient
+from Databases.Mongo.MongoClient import MongoClient
 from Components.Monitors.WorkflowCloseoutMonitor import WorkflowCloseoutMonitor
